Interpolation Dem
-Demonstration of the interpolation class
